'medianColor' no longer affects current drawing context if not set
Only sets the stroke style if there is a color set (e.g. not null) else it uses the current border color. This also rearranges the drawing order so that the box outline is drawn on top of the median line, so that you don't end up with the median line cutting into the border.